Finalized main routes, functions, etc.

+++ app/Http/Controllers/GroupController.php
@@ -5,7 +5,10 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Auth;
+use App\Rules\IsMember;
+
 use App\Http\Resources\Group as GroupResource;
+use App\Http\Resources\Member as MemberResource;
 use App\User;
 use App\Group;
 
@@ -50,20 +53,6 @@ class GroupController extends Controller
     {
         return new GroupResource($group);
     }
-    
-    /* public function show(Group $group)
-    {
-        $user = Auth::guard('api')->user();
-        if($group->members->contains($user)){
-            return new JsonResource([
-                    'group_id' => $group->id,
-                    'group_name' => $group->name,
-                    'member' => new MemberResource($group->members->find($user))
-                ]);
-        } else {
-            return response()->json(['error' => 'This user is not a member of this group.'], 400);
-        }
-    } */
 
     public function store(Request $request)
     {
@@ -87,16 +76,35 @@ class GroupController extends Controller
         if($validator->fails()){
             return response()->json(['error' => $validator->errors()], 400);
         }
-        $group->update($request->all());
-        return response()->json(new GroupResource($group), 200);
+        $user = Auth::guard('api')->user();
+        $member = $group->members->find($user);
+        if($member == null){
+            abort(400, 'User is not a member of this group.');
+        }
+        if($member->member_data->is_admin){
+            $group->update($request->all());
+            return response()->json(new GroupResource($group), 200);
+        } else {
+            return response()->json(['error' => 'User is not an admin'], 400);
+        }
     }
 
     public function delete(Group $group)
     {
-        $group->members()->detach($group->members);
-        $group->delete();
+        $user = Auth::guard('api')->user();
+        $member = $group->members->find($user);
+        if($member == null){
+            abort(400, 'User is not a member of this group.');
+        }
+        if($member->member_data->is_admin){
+            $group->members()->detach($group->members);
+            $group->delete();
+    
+            return response()->json(null, 204);
+        } else {
+            return response()->json(['error' => 'User is not an admin'], 400);
+        }
 
-        return response()->json(null, 204);
     }
 
     /* Members */
@@ -106,27 +114,42 @@ class GroupController extends Controller
         if($group->members->count()==20){
             return response()->json(['error' => 'Group member limit reached'], 400);
         }
-        $group->members()->attach($user, [
-            'nickname' => $request->nickname ?? null,
-            'is_admin' => $group->members()->count() == 0 //set to true on first member
-        ]);
+        if($group->members->find($user) == null){
+            $group->members()->attach($user, [
+                'nickname' => $request->nickname ?? null,
+                'is_admin' => $group->members()->count() == 0 //set to true on first member
+            ]);
+        } else {
+            return response()->json(['error' => 'The user is already a member in this group'], 400);
+        }
         
         return response()->json(null, 204);
     }
 
     public function updateMember(Group $group, Request $request)
     {
+        $user = Auth::guard('api')->user();
+        $member = $group->members->find($user);
+        if($member == null){
+            abort(400, 'User is not a member of this group.');
+        }
         $validator = Validator::make($request->all(), [
+            'member_id' => ['required','exists:users,id', new IsMember($group->id)],
             'nickname' => 'string|min:3',
             'is_admin' => 'boolean'
         ]);
         if($validator->fails()){
             return response()->json(['error' => $validator->errors()], 400);
         }
-
-        $user = Auth::guard('api')->user();
-        $member = $group->members->findOrFail($user);
-        $member->member_data->update($request->all());
+        if($member == $group->members->find($request->member_id)){
+            $member->member_data->update($request->except('member_id'));
+        } else {
+            if($member->member_data->is_admin){
+                $group->members->find($request->member_id)->member_data->update($request->except('member_id'));
+            } else {
+                return response()->json(['error' => 'User is not admin'], 400);
+            }
+        }
 
         return response()->json(null, 204);
